@@ -1,6 +1,6 @@
 module github.com/kubevirt/hyperconverged-cluster-operator/tools/release-notes
 
-go 1.18
+go 1.19
 
 require (
 	github.com/golang/glog v0.0.0-20160126235308-23def4e6c14b
